requirements.txt
asgiref==3.5.2
cffi==1.15.1
Django==4.1.3
hunspell==0.5.5
lxml==4.9.1
pkg_resources==0.0.0
pycparser==2.21
PyStemmer==2.0.1
snowballstemmer==2.2.0
sqlitefts==0.5.1
sqlparse==0.4.3
